Fix bug del Store de presupuestos al agregar los datos

+++ app/model/Presupuesto.js
@@ -3,13 +3,14 @@ Ext.define('myMoney.model.Presupuesto', {
 	requires: 'Ext.data.identifier.Uuid',
 	
 	config: {
+		
 		identifier: {
 			type: 'uuid'
 		},
 		
 		fields: [
 			{name: 'name', type: 'string'},
-			{name: 'monto', type: 'double'},
+			{name: 'monto', type: 'float'},
 		]
 	}
 });
